revert movement of SigmaParser tests
